requirements.txt
pywebview
flask
flask-cors
flask-socketio
eventlet
simple-websocket